backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-magic>=0.4.27
PyPDF2>=3.0.0
pandas>=1.3.0
nltk>=3.6.3
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=0.19.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.29.3
stripe==7.6.0
